# backend/guia_cerrado/docs/swagger.json
{
...
  "info": {
    "description": "Documentação da API do Guia Cerrado",
    "termsOfService": "/tos",
    "title": "Guia Cerrado API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/cities-by-state": {
      "get": {
        "description": "Endpoint para consultar todas as cidades distintas associadas a um estado específico nos registros de preços, ordenadas alfabeticamente.\n",
        "parameters": [
          {
            "description": "Sigla do estado",
            "example": "GO",
            "in": "query",
            "name": "state",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de cidades retornada com sucesso",
            "examples": {
              "application/json": [
                "Goiânia",
                "Jataí",
                "Rio Verde"
              ]
            },
            "schema": {
              "items": {
                "example": "Rio Verde",
                "type": "string"
              },
              "type": "array"
            }
          },
          "400": {
            "description": "Parâmetro obrigatório não informado",
            "examples": {
              "application/json": {
                "error": "state parameter is required"
              }
            },
            "schema": {
              "properties": {
                "error": {
                  "example": "state parameter is required",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Lista todas as cidades de um estado específico",
        "tags": [
          "Pricing"
        ]
      }
    },
...
    "/price-stats": {
      "get": {
        "description": "Endpoint para consultar estatísticas agregadas de preços por cultura, incluindo valor mínimo, máximo e médio. Os resultados podem ser filtrados opcionalmente por estado e cidade.\n",
        "parameters": [
          {
            "description": "Sigla do estado para filtrar os resultados",
            "example": "GO",
            "in": "query",
            "name": "state",
            "required": false,
            "type": "string"
          },
          {
            "description": "Nome da cidade para filtrar os resultados",
            "example": "Rio Verde",
            "in": "query",
            "name": "city",
            "required": false,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Estatísticas de preços retornadas com sucesso",
            "examples": {
              "application/json": [
                {
                  "avg_price": 133.2,
                  "culture_id": 1,
                  "culture_name": "Soja",
                  "max_price": 145.9,
                  "min_price": 120.5
                },
                {
                  "avg_price": 95.7,
                  "culture_id": 2,
                  "culture_name": "Milho",
                  "max_price": 102.4,
                  "min_price": 89.0
                }
              ]
            },
            "schema": {
              "items": {
                "properties": {
                  "avg_price": {
                    "example": 133.2,
                    "format": "float",
                    "type": "number"
                  },
                  "culture_id": {
                    "example": 1,
                    "type": "integer"
                  },
                  "culture_name": {
                    "example": "Soja",
                    "type": "string"
                  },
                  "max_price": {
                    "example": 145.9,
                    "format": "float",
                    "type": "number"
                  },
                  "min_price": {
                    "example": 120.5,
                    "format": "float",
                    "type": "number"
                  }
                },
                "required": [
                  "culture_id",
                  "culture_name",
                  "min_price",
                  "max_price",
                  "avg_price"
                ],
                "type": "object"
              },
              "type": "array"
            }
          }
        },
        "summary": "Retorna estatísticas de preços por cultura",
        "tags": [
          "Pricing"
        ]
      }
    },
    "/prices": {
      "get": {
        "description": "Endpoint para listar os registros de preço associados a uma cultura agrícola específica, informada pelo parâmetro culture_id.\n",
        "parameters": [
          {
            "description": "ID da cultura agrícola",
            "example": 1,
            "in": "query",
            "name": "culture_id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de preços retornada com sucesso",
            "examples": {
              "application/json": [
                {
                  "city": "Rio Verde",
                  "culture_id": 1,
                  "end_date": "Fri, 15 May 2026 00:00:00 GMT",
                  "id": 10,
                  "price": 135.75,
                  "start_date": "Fri, 01 May 2026 00:00:00 GMT",
                  "state": "GO"
                },
                {
                  "city": "Jataí",
                  "culture_id": 1,
                  "end_date": "Fri, 15 May 2026 00:00:00 GMT",
                  "id": 11,
                  "price": 132.4,
                  "start_date": "Fri, 01 May 2026 00:00:00 GMT",
                  "state": "GO"
                }
              ]
            },
            "schema": {
              "items": {
                "properties": {
                  "city": {
                    "example": "Rio Verde",
                    "type": "string"
                  },
                  "culture_id": {
                    "example": 1,
                    "type": "integer"
                  },
                  "end_date": {
                    "example": "Fri, 15 May 2026 00:00:00 GMT",
                    "format": "date",
                    "type": "string"
                  },
                  "id": {
                    "example": 10,
                    "type": "integer"
                  },
                  "price": {
                    "example": 135.75,
                    "format": "float",
                    "type": "number"
                  },
                  "start_date": {
                    "example": "Fri, 01 May 2026 00:00:00 GMT",
                    "format": "date",
                    "type": "string"
                  },
                  "state": {
                    "example": "GO",
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "culture_id",
                  "city",
                  "state",
                  "price",
                  "start_date",
                  "end_date"
                ],
                "type": "object"
              },
              "type": "array"
            }
          },
          "400": {
            "description": "Parâmetro obrigatório não informado",
            "examples": {
              "application/json": {
                "error": "culture_id parameter is required"
              }
            },
            "schema": {
              "properties": {
                "error": {
                  "example": "culture_id parameter is required",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Retorna os registros de preço de uma cultura específica",
        "tags": [
          "Pricing"
        ]
      }
    },
    "/prices-by-location": {
      "get": {
        "description": "Endpoint para listar os registros de preço de todas as culturas agrícolas para uma cidade e estado específicos.\n",
        "parameters": [
          {
            "description": "Sigla do estado",
            "example": "GO",
            "in": "query",
            "name": "state",
            "required": true,
            "type": "string"
          },
          {
            "description": "Nome da cidade",
            "example": "Rio Verde",
            "in": "query",
            "name": "city",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de preços da localização retornada com sucesso",
            "examples": {
              "application/json": [
                {
                  "city": "Rio Verde",
                  "culture_id": 2,
                  "end_date": "Fri, 15 May 2026 00:00:00 GMT",
                  "id": 21,
                  "price": 98.3,
                  "start_date": "Fri, 01 May 2026 00:00:00 GMT",
                  "state": "GO"
                },
                {
                  "city": "Rio Verde",
                  "culture_id": 5,
                  "end_date": "Fri, 15 May 2026 00:00:00 GMT",
                  "id": 22,
                  "price": 145.9,
                  "start_date": "Fri, 01 May 2026 00:00:00 GMT",
                  "state": "GO"
                }
              ]
            },
            "schema": {
              "items": {
                "properties": {
                  "city": {
                    "example": "Rio Verde",
                    "type": "string"
                  },
                  "culture_id": {
                    "example": 2,
                    "type": "integer"
                  },
                  "end_date": {
                    "example": "Fri, 15 May 2026 00:00:00 GMT",
                    "format": "date",
                    "type": "string"
                  },
                  "id": {
                    "example": 21,
                    "type": "integer"
                  },
                  "price": {
                    "example": 98.3,
                    "format": "float",
                    "type": "number"
                  },
                  "start_date": {
                    "example": "Fri, 01 May 2026 00:00:00 GMT",
                    "format": "date",
                    "type": "string"
                  },
                  "state": {
                    "example": "GO",
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "culture_id",
                  "city",
                  "state",
                  "price",
                  "start_date",
                  "end_date"
                ],
                "type": "object"
              },
              "type": "array"
            }
          },
          "400": {
            "description": "Parâmetros obrigatórios não informados",
            "examples": {
              "application/json": {
                "error": "Both state and city parameters are required"
              }
            },
            "schema": {
              "properties": {
                "error": {
                  "example": "Both state and city parameters are required",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Retorna preços de todas as culturas para uma localização específica",
        "tags": [
          "Pricing"
        ]
      }
    },
    "/recent-prices": {
      "get": {
        "description": "Endpoint para consultar os registros de preço mais recentes para cada combinação de cultura e cidade dentro de um estado específico, com base na maior data final disponível.\n",
        "parameters": [
          {
            "description": "Sigla do estado usado no filtro",
            "example": "GO",
            "in": "query",
            "name": "state",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Lista de registros mais recentes retornada com sucesso",
            "examples": {
              "application/json": [
                {
                  "city": "Rio Verde",
                  "culture_id": 1,
                  "end_date": "Fri, 15 May 2026 00:00:00 GMT",
                  "id": 101,
                  "price": 138.2,
                  "start_date": "Fri, 01 May 2026 00:00:00 GMT",
                  "state": "GO"
                },
                {
                  "city": "Jataí",
                  "culture_id": 2,
                  "end_date": "Fri, 15 May 2026 00:00:00 GMT",
                  "id": 102,
                  "price": 97.4,
                  "start_date": "Fri, 01 May 2026 00:00:00 GMT",
                  "state": "GO"
                }
              ]
            },
            "schema": {
              "items": {
                "properties": {
                  "city": {
                    "example": "Rio Verde",
                    "type": "string"
                  },
                  "culture_id": {
                    "example": 1,
                    "type": "integer"
                  },
                  "end_date": {
                    "example": "Fri, 15 May 2026 00:00:00 GMT",
                    "format": "date",
                    "type": "string"
                  },
                  "id": {
                    "example": 101,
                    "type": "integer"
                  },
                  "price": {
                    "example": 138.2,
                    "format": "float",
                    "type": "number"
                  },
                  "start_date": {
                    "example": "Fri, 01 May 2026 00:00:00 GMT",
                    "format": "date",
                    "type": "string"
                  },
                  "state": {
                    "example": "GO",
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "culture_id",
                  "city",
                  "state",
                  "price",
                  "start_date",
                  "end_date"
                ],
                "type": "object"
              },
              "type": "array"
            }
          },
          "400": {
            "description": "Parâmetro obrigatório não informado",
            "examples": {
              "application/json": {
                "error": "state parameter is required"
              }
            },
            "schema": {
              "properties": {
                "error": {
                  "example": "state parameter is required",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "summary": "Retorna os registros de preço mais recentes por cultura e localização",
        "tags": [
          "Pricing"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}